+++ src/full/Agda/TypeChecking/ReconstructParameters.hs
@@ -15,7 +15,6 @@ import Agda.Syntax.Internal.Generic
 
 import Agda.TypeChecking.Monad
 import Agda.TypeChecking.CheckInternal
-import Agda.TypeChecking.Irrelevance
 import Agda.TypeChecking.ProjectionLike
 import Agda.TypeChecking.Substitute
 import Agda.TypeChecking.Reduce
